pyproject.toml
[project]
name = "searchless-chess"
version = "0.1.0"
description = "Amortized Planning with Large-Scale Transformers: A Case Study on Chess"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "absl-py",
    "apache-beam",
    "chess",
    "chex",
    "dm-haiku",
    "grain-nightly",
    "jax",
    "jaxtyping",
    "jupyter",
    "numpy",
    "optax",
    "orbax-checkpoint",
    "pandas",
    "scipy",
    "typing-extensions",
]

[project.optional-dependencies]
cuda = [
    "jax[cuda12]",
]
antichess = [
    "datasets",
    "polars",
    "tqdm",
    "pyarrow",
]

[tool.uv]
dev-dependencies = []
